feat: Add tooltip to list type switch and move response model
This commit introduces two main changes:
1.  Adds a tooltip to the "Anime/Manga" switch in the `MyListScreen`.
2.  Moves the `UpdatedAnimeResponse` data class to its own file within the `com.branwen.mal.models.remote` package.

**Key Changes:**

*   **`MyListScreen.kt`**:
    *   Wrapped the `Text` component (displaying "A" or "M") inside the `Switch` with a `TooltipBox`.
    *   The tooltip displays "List Type" when the user interacts with the switch.
    *   Utilized `TooltipDefaults.rememberPlainTooltipPositionProvider()` for tooltip positioning and `rememberTooltipState()` for managing its state.

*   **`UpdatedAnimeResponse.kt`**:
    *   Created a new file `app/src/main/java/com/branwen/mal/models/remote/UpdatedAnimeResponse.kt`.
    *   Moved the `UpdatedAnimeResponse` data class from `AnimeRepository.kt` to this new file.

*   **`AnimeRepository.kt`**:
    *   Removed the `UpdatedAnimeResponse` data class definition as it's now in its own file.
    *   Updated imports to reflect the new location of `UpdatedAnimeResponse`.

*   **`AnimeRemoteDataSource.kt`**:
    *   Updated imports to reflect the new location of `UpdatedAnimeResponse`.

This improves user experience by providing context for the list type switch and enhances code organization by placing the response model in a more appropriate location.
