+++ app/src/main/java/com/branwen/mal/data/repo/AnimeRepository.kt
@@ -4,8 +4,6 @@ import com.branwen.mal.data.local.AnimeLocalDataSource
 import com.branwen.mal.data.remote.AnimeRemoteDataSource
 import com.branwen.mal.models.AnimeNode
 import com.branwen.mal.models.domain.MyAnimeListItem
-import com.squareup.moshi.Json
-import com.squareup.moshi.JsonClass
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.emitAll
 import kotlinx.coroutines.flow.firstOrNull
@@ -70,17 +68,3 @@ class AnimeRepository(
         }
     }
 }
-
-@JsonClass(generateAdapter = true)
-data class UpdatedAnimeResponse(
-    val status: String,
-    val score: Int,
-    @Json(name = "num_episodes_watched") val numEpisodesWatched: Int,
-    @Json(name = "is_rewatching") val isRewatching: Boolean,
-    @Json(name = "updated_at") val updatedAt: String,
-    val priority: Int,
-    @Json(name = "num_times_rewatched") val numTimesRewatched: Int,
-    @Json(name = "rewatch_value") val rewatchValue: Int,
-    val tags: List<String>,
-    val comments: String
-)
